+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Module ReMOVE Lite
     - LEDs for RUN and RESET turn red when using PHASE-input
     - OUT-port can be used for monitoring while recording
+    - OUT-port bypasses IN-port when selecting an empty sequence
     - Added play mode "Sequence Loop"
     - Fixed bug when saving sequences with lots of constant values (compression-bug)